move single picker out
